Tests: move tests to a "tests" subfolder
In order to make folder scanning easier.
