+++ src/container/tests/extensionApiSpec.js
@@ -2,9 +2,9 @@
 /* eslint-env mocha */
 
 import * as sinon from 'sinon';
-import { containerDoubles } from '../../test/doubles';
-import { ArgError } from './errors';
-import ExtensionApi from './ExtensionApi';
+import { containerDoubles } from '../../../test/doubles';
+import { ArgError } from '../errors';
+import ExtensionApi from '../ExtensionApi';
 
 describe('ExtensionApi', () => {
