+++ src/extensions/tests/commonArgResolverSpec.js
@@ -1,9 +1,9 @@
 /* eslint no-unused-expressions: 0, max-nested-callbacks: 0 */
 /* eslint-env mocha */
 
-import { _ } from '../globals';
-import { containerDoubles } from '../../test/doubles';
-import CommonArgResolver from './CommonArgResolver';
+import { _ } from '../../globals';
+import { containerDoubles } from '../../../test/doubles';
+import CommonArgResolver from '../CommonArgResolver';
 
 describe('CommonArgResolver', () => {
