+++ src/extensions/tests/serviceArgResolverSpec.js
@@ -1,9 +1,9 @@
 /* eslint no-unused-expressions: 0, max-nested-callbacks: 0 */
 /* eslint-env mocha */
 
-import { containerDoubles } from '../../test/doubles';
-import { addSpecsForCanResolveArgMatchingPrefix } from '../../test/argResolvers';
-import ServiceArgResolver from './ServiceArgResolver';
+import { containerDoubles } from '../../../test/doubles';
+import { addSpecsForCanResolveArgMatchingPrefix } from '../../../test/argResolvers';
+import ServiceArgResolver from '../ServiceArgResolver';
 
 describe('ServiceArgResolver', () => {
