+++ src/extensions/tests/subscriptionManagerExtensionSpec.js
@@ -1,10 +1,10 @@
 /* eslint no-unused-expressions: 0, max-nested-callbacks: 0 */
 /* eslint-env mocha */
 
-import { _ } from '../globals';
+import { _ } from '../../globals';
 import * as sinon from 'sinon';
-import { containerDoubles } from '../../test/doubles';
-import SubscriptionManagerExtension from './SubscriptionManagerExtension';
+import { containerDoubles } from '../../../test/doubles';
+import SubscriptionManagerExtension from '../SubscriptionManagerExtension';
 
 describe('SubscriptionManagerExtension', () => {
